Rename document files
