+++ imports/startup/client/routes.js
@@ -6,9 +6,9 @@ import { Router, Route, IndexRoute, browserHistory } from 'react-router';
 import { Meteor } from 'meteor/meteor';
 import App from '../../ui/layouts/App.js';
 import Documents from '../../ui/pages/documents/Documents.js';
-import NewDocument from '../../ui/pages/documents/NewDocument.js';
-import EditDocument from '../../ui/containers/documents/EditDocument.js';
-import ViewDocument from '../../ui/containers/documents/ViewDocument.js';
+import NewDocument from '../../ui/pages/documents/New.js';
+import EditDocument from '../../ui/containers/documents/Edit.js';
+import ViewDocument from '../../ui/containers/documents/View.js';
 import Index from '../../ui/pages/Index.js';
 import Login from '../../ui/pages/Login.js';
 import NotFound from '../../ui/pages/NotFound.js';
