+++ imports/ui/pages/documents/Documents.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Row, Col, Button } from 'react-bootstrap';
-import DocumentsList from '../../containers/documents/DocumentsList.js';
+import DocumentsList from '../../containers/documents/List.js';
 
 const Documents = () => (
   <div className="Documents">
