+++ imports/ui/pages/documents/New.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import DocumentEditor from '../../components/documents/DocumentEditor.js';
+import DocumentEditor from '../../components/documents/Editor.js';
 
 const NewDocument = () => (
   <div className="NewDocument">
